move safe_permute to utils
