+++ src/compressed_tensors/quantization/lifecycle/forward.py
@@ -17,7 +17,6 @@ from math import ceil
 from typing import Optional
 
 import torch
-from compressed_tensors.quantization.lifecycle.helpers import safe_permute
 from compressed_tensors.quantization.observers.helpers import calculate_range
 from compressed_tensors.quantization.quant_args import (
     QuantizationArgs,
@@ -26,8 +25,7 @@ from compressed_tensors.quantization.quant_args import (
 )
 from compressed_tensors.quantization.quant_config import QuantizationStatus
 from compressed_tensors.quantization.quant_scheme import QuantizationScheme
-from compressed_tensors.utils import update_parameter_data
-from compressed_tensors.quantization.lifecycle.helpers import safe_permute
+from compressed_tensors.utils import update_parameter_data, safe_permute
 from torch.nn import Module
