+++ src/compressed_tensors/utils/__init__.py
@@ -15,6 +15,7 @@
 
 from .helpers import *
 from .offload import *
+from .permute import *
 from .permutations_24 import *
 from .safetensors_load import *
 from .semi_structured_conversions import *
